Move function for getting float vectors
